Make the reaper container respect `DOCKER_HOST`
Problem: The reaper container currently expects a docker socket to be
available at `/var/run/docker.sock`, but this is not always the
case. Especially with rootless docker, which usually puts the socket
at `$XDG_RUNTIME_DIR/docker.sock` and then exports it as
`DOCKER_HOST=unix://$XDG_RUNTIME_DIR/docker.sock`. In this case, the
reaper container will not be able use docker.

Solution: Check if the `DOCKER_HOST` environment variable sets the
docker socket location to something other than the default.